Перенесет действия из app-process d app-data
